+++ src/recv_ring_stream.h
@@ -124,6 +124,11 @@ void ring_stream<Ringbuffer>::heap_ready(live_heap &&h)
                      h.get_cnt());
         }
     }
+    else
+    {
+        log_warning("dropped incomplete heap %d (%d/%d bytes of payload)",
+                    h.get_received_length(), h.get_heap_length());
+    }
 }
 
 template<typename Ringbuffer>